shape ops test fixes
- fix rand for Android
- build unit test on linux
- use atomic inc in test count
- add casting for Android

git-svn-id: http://skia.googlecode.com/svn/trunk@8610 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ tests/PathOpsQuadLineIntersectionThreadedTest.cpp
@@ -81,7 +81,8 @@ static void testQuadLineIntersectMain(PathOpsThreadState* data)
     int by = state.fB >> 2;
     int cx = state.fC & 0x03;
     int cy = state.fC >> 2;
-    SkDQuad quad = {{{ax, ay}, {bx, by}, {cx, cy}}};
+    SkDQuad quad = {{{(double) ax, (double) ay}, {(double) bx, (double) by},
+            {(double) cx, (double) cy}}};
     SkReduceOrder reducer;
     int order = reducer.reduce(quad, SkReduceOrder::kFill_Style);
     if (order < 3) {
